warnings removed
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@1521 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/wall1/w1_mat_dam_mp.c
@@ -40,8 +40,8 @@ void w1_mat_dam_mp(DOUBLE     youngs,
                    INT        ip,
                    DOUBLE    *stress,
                    DOUBLE   **D,
-                   DOUBLE     istore,
-                   DOUBLE     newval) 
+                   INT        istore,
+                   INT        newval) 
 {
 INT i,j,k;
 INT yip;
